requirements.txt
numpy
scikit-learn
scipy
tqdm
torch
torchvision
pyyaml
dotenv
pandas
matplotlib